pyproject.toml
[project]
name = "RPG_API"
version = "1.0.0"
dependencies = [
  "Django==4.2.*",
  "djangorestframework",
  "django-environ",
  "psycopg2"
]

[tool.setuptools]
py-modules = []
